requirements.txt
coverage==4.1
defusedxml==0.4.1
Django==1.9.6
django-braces==1.9.0
django-cors-headers==1.1.0
django-qiniu-storage==2.3.0
django-redis==4.4.3
django-taggit==0.21.3
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
mysqlclient==1.3.7
oauthlib==1.0.3
PyJWT==1.4.0
python-social-auth==0.2.21
python3-openid==3.0.10
qiniu==7.0.7
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
rest-social-auth==0.4.4
shortuuid==0.4.3
six==1.10.0
sqlparse==0.1.19
uWSGI==2.0.13.1
